update sing-a10l.pl
Former-commit-id: 5293655d074c10e0b4f882ee373d94473d014408

+++ Workfile/sign-image/sign-a10l.pl
@@ -5,19 +5,29 @@ use 5.010;
 
 my $exp_script = "sg.exp";
 my $odir = "test-sign";
+my $ofile = "signed-a10l-img-DATE.zip";
 my $intermediate = "signed-target-files.zip";
 my $certdir = "/home/rasmus/duke";
+my $delete_intermediate = 0;
 
-sub sign_first()
+sub sign_first($)
 {
-    my $REPLACE_FILE=`find ./dist/ -name 'tostab12AL-target_files*zip'`;
-    $REPLACE_FILE =~ s/[\r\n]//;
-    #say $REPLACE_FILE;
+	my $file = shift;
+    my $REPLACE_FILE;
+
+	if ( -e $file ) {
+		say "=====> file found...";
+		$REPLACE_FILE = $file;
+	} else {
+        $REPLACE_FILE = `find ./dist/ -name 'tostab12AL-target_files*zip'`;
+        $REPLACE_FILE =~ s/[\r\n]//;
+	}
+    say "=====> from file: " . $REPLACE_FILE;
 
     my $script=<<SCRIPTEOL;
 #!/usr/bin/expect
 
-spawn ./build/tools/releasetools/sign_target_files_apks -d $certdir -e motiv_theme_1_Strong.apk= -e motiv_theme_2_Medium.apk= -e motiv_theme_3_Soft.apk= -e motiv_theme_4_None.apk= REPLACE_ME $odir/$intermediate
+spawn ./build/tools/releasetools/sign_target_files_apks -d $certdir REPLACE_ME $odir/$intermediate
 
 expect "Enter password for $certdir/media key>"
 send -- "28689155\\r"
@@ -42,19 +52,37 @@ SCRIPTEOL
 
 sub sign_second()
 {
-    my $cmd = "./build/tools/releasetools/img_from_target_files $odir/$intermediate $odir/signed-img.zip";
+    my $cmd = "./build/tools/releasetools/img_from_target_files $odir/$intermediate $odir/$ofile";
 	print $cmd,"\n";
     system $cmd;
 }
 
 sub main()
 {
-    sign_first();
-    print "run expect...";
+    my $date = `date +%y%m%d-%H%M`;
+	$date =~ s/[\r\n]//;
+#	say $date;
+	$ofile =~ s/DATE/$date/;
+#	say $ofile;
+
+	unless ( -d $odir ) {
+        mkdir $odir;
+    }
+
+	my $ifile = $ARGV[0] // "";
+    sign_first($ifile);
+    say "=====> run expect...";
     system "expect $exp_script";
-    unlink $exp_script;
+	if ($delete_intermediate) {
+        unlink $exp_script;
+    }
+
     sign_second();
-	unlink "$odir/$intermediate";
+    if ($delete_intermediate) {
+	    unlink "$odir/$intermediate";
+    }
+
+	say "=====> output file at: $odir/$ofile";
 }
 
 main;
